Rename test_case_utils to case_utils
buck gets confused with utility module name with `test_` prefix.
